[refactor] 그래프 반응형 넓이 이벤트 적용 + index.js로 파일 이름 변경
